+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Added
 
 - New layout type "screenshot-only" that enables to render a screenshot without a caption. The screenshot is centered inside the original image at the desired size factor and may be decorated with a shadow and background as usual.
+- Now also screenshots in JPEG format are supported, additionally to PNG. Note that the output is still always in PNG format.
 
 ### Changed